make all toolbar items respect settings, and refactor the editor component
